@@ -1,12 +1,13 @@
 import commonjs from 'rollup-plugin-commonjs';
 import resolve from 'rollup-plugin-node-resolve';
-// import typescript from 'rollup-plugin-typescript2';
+import typescript from 'rollup-plugin-typescript2';
+import babel from 'rollup-plugin-babel';
 
 import pkg from './package.json';
 
 const external = id => !id.startsWith('.') && !id.startsWith('/');
 
-const input = 'src/index.jsx';
+const input = 'src/index.js';
 
 const buildEs = {
   input,
@@ -25,7 +26,15 @@ const buildEs = {
       sourcemap: true
     }
   ],
-  plugins: [resolve(), commonjs()]
+  plugins: [
+    resolve(),
+    babel(),
+    typescript({
+      rollupCommonJSResolveHack: true,
+      clean: true
+    }),
+    commonjs()
+  ]
 };
 
 export default buildEs;
